move import project service into FwLiteShared
