+++ backend/FwLite/FwLiteShared/Auth/AuthHelpers.cs
@@ -1,4 +1,5 @@
 using System.Net.Http.Headers;
+using FwLiteShared.Projects;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
