+++ backend/FwLite/FwLiteShared/FwLiteSharedKernel.cs
@@ -1,5 +1,7 @@
 ﻿using FwDataMiniLcmBridge;
+using FwLiteProjectSync;
 using FwLiteShared.Auth;
+using FwLiteShared.Projects;
 using FwLiteShared.Sync;
 using LcmCrdt;
 using Microsoft.Extensions.DependencyInjection;
@@ -15,7 +17,9 @@ public static class FwLiteSharedKernel
         services.AddAuthHelpers(environment);
         services.AddLcmCrdtClient();
         services.AddFwDataBridge();
+        services.AddFwLiteProjectSync();
 
+        services.AddSingleton<ImportFwdataService>();
         services.AddScoped<SyncService>();
         services.AddSingleton<LexboxProjectService>();
         services.AddSingleton<ChangeEventBus>();
