+++ backend/FwLite/LocalWebApp/Hubs/CrdtMiniLcmApiHub.cs
@@ -1,4 +1,5 @@
 ﻿using FwLiteShared;
+using FwLiteShared.Projects;
 using FwLiteShared.Sync;
 using LcmCrdt;
 using LcmCrdt.Data;
