+++ backend/FwLite/LocalWebApp/Routes/ImportRoutes.cs
@@ -1,4 +1,5 @@
-﻿ using SIL.Harmony.Db;
+﻿ using FwLiteShared.Projects;
+ using SIL.Harmony.Db;
  using LocalWebApp.Services;
  using Microsoft.OpenApi.Models;
 using MiniLcm;
